+++ R/enviro-par.R
@@ -11,12 +11,12 @@
 #' @export
 
 enviro_par <- function(.x) {
-  
+
   which <- "enviro"
-  nms <- tealeaves::parameter_names(which)
-  
+  nms <- parameter_names(which)
+
   stopifnot(is.list(.x))
-  
+
   if (!all(nms %in% names(.x))) {
     nms[!(nms %in% names(.x))] %>%
       stringr::str_c(collapse = ", ") %>%
@@ -24,9 +24,9 @@ enviro_par <- function(.x) {
                  x = ., which = which) %>%
       stop()
   }
-  
+
   .x %<>% magrittr::extract(nms)
-  
+
   # Set units ----
   .x$P %<>% set_units(kPa)
   .x$r %<>% set_units()
@@ -34,7 +34,7 @@ enviro_par <- function(.x) {
   .x$S_sw %<>% set_units(W / m ^ 2)
   .x$T_air %<>% set_units(K)
   .x$wind %<>% set_units(m / s)
-  
+
   # Check values ----
   stopifnot(.x$P >= set_units(0, kPa))
   stopifnot(.x$r >= set_units(0) & .x$r <= set_units(1))
@@ -48,23 +48,23 @@ enviro_par <- function(.x) {
     message(
       '\ntealeaves (>= 1.0.2) will require users provide a T_sky value or function\nto calculate T_sky from other parameters.\n\nFor back-compatibility, if T_sky is not provided, this warning will appear\nand the default function used in tealeaves (< 1.0.2) will be applied.\n\nSee more details in vignette("parameter-functions")'
     )
-    
-    .x$T_sky <- function(pars) {pars$T_air - set_units(20, K) * 
+
+    .x$T_sky <- function(pars) {pars$T_air - set_units(20, K) *
         pars$S_sw / set_units(1000, W / m ^ 2)}
-      
+
   } else {
-    
+
     stopifnot(is.function(.x$T_sky) | is.double(.x$T_sky))
-    
+
     if (is.double(.x$T_sky)) {
-      
+
       .x$T_sky %<>% set_units(K)
-    
+
     }
-    
+
   }
-  
+
   structure(.x, class = c(stringr::str_c(which, "_par"), "list"))
-  
+
 }